refactor: rework case strikeout scheduler (CMC-1172)
