+++ ccd-definition/AuthorisationCaseField/solicitor.json
@@ -614,5 +614,12 @@
     "CaseFieldID": "respondent2OrganisationPolicy",
     "UserRole": "caseworker-civil-solicitor",
     "CRUD": "CRU"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "CaseTypeID": "UNSPECIFIED_CLAIMS",
+    "CaseFieldID": "applicantSolicitor1ClaimStrikeOutDeadlineToRespondentSolicitor1",
+    "UserRole": "caseworker-civil-solicitor",
+    "CRUD": "CRUD"
   }
 ]
