+++ ccd-definition/AuthorisationCaseField/systemUpdate.json
@@ -331,7 +331,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "CaseFieldID": "applicantSolicitorSecondResponseDeadlineToRespondentSolicitor1",
+    "CaseFieldID": "applicantSolicitor1ClaimStrikeOutDeadlineToRespondentSolicitor1",
     "UserRole": "caseworker-civil-systemupdate",
     "CRUD": "CRUD"
   },
