+++ ccd-definition/AuthorisationCaseState/systemUpdate.json
@@ -33,5 +33,12 @@
     "CaseStateID": "PROCEEDS_WITH_OFFLINE_JOURNEY",
     "UserRole": "caseworker-civil-systemupdate",
     "CRUD": "CRUD"
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "CaseTypeID": "UNSPECIFIED_CLAIMS",
+    "CaseStateID": "CASE_STRUCK_OUT",
+    "UserRole": "caseworker-civil-systemupdate",
+    "CRUD": "CRUD"
   }
 ]
