+++ ccd-definition/CaseEvent/User/UserEvents.json
@@ -82,23 +82,10 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "ID": "MOVE_TO_STAYED",
-    "Name": "Move case to stayed",
-    "Description": "Moves a case to stayed state",
-    "DisplayOrder": 5,
-    "PreConditionState(s)": "PENDING_CASE_ISSUED;CREATED",
-    "PostConditionState": "STAYED",
-    "SecurityClassification": "Public",
-    "ShowSummary": "N",
-    "ShowEventNotes": "N"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "UNSPECIFIED_CLAIMS",
     "ID": "ACKNOWLEDGE_SERVICE",
     "Name": "Acknowledge service",
     "Description": "Defendant solicitor is acknowledging service",
-    "DisplayOrder": 6,
+    "DisplayOrder": 5,
     "PreConditionState(s)": "PENDING_CASE_ISSUED;CREATED",
     "PostConditionState": "*",
     "SecurityClassification": "Public",
@@ -118,7 +105,7 @@
     "ID": "ADD_DEFENDANT_LITIGATION_FRIEND",
     "Name": "Add litigation friend",
     "Description": "Add litigation friend",
-    "DisplayOrder": 7,
+    "DisplayOrder": 6,
     "PreConditionState(s)": "PENDING_CASE_ISSUED;CREATED;AWAITING_CLAIMANT_INTENTION",
     "PostConditionState": "*",
     "SecurityClassification": "Public",
@@ -133,7 +120,7 @@
     "ID": "DEFENDANT_RESPONSE",
     "Name": "Respond to claim",
     "Description": "Defendant response to claim",
-    "DisplayOrder": 8,
+    "DisplayOrder": 7,
     "PreConditionState(s)": "PENDING_CASE_ISSUED;CREATED",
     "PostConditionState": "AWAITING_CLAIMANT_INTENTION",
     "SecurityClassification": "Public",
@@ -153,7 +140,7 @@
     "ID": "CLAIMANT_RESPONSE",
     "Name": "View and respond to defence",
     "Description": "View and respond to defendant",
-    "DisplayOrder": 9,
+    "DisplayOrder": 8,
     "PreConditionState(s)": "AWAITING_CLAIMANT_INTENTION",
     "PostConditionState": "STAYED",
     "SecurityClassification": "Public",
@@ -173,9 +160,9 @@
     "ID": "MOVE_CLAIM_TO_STRUCK_OUT",
     "Name": "Strike out claim",
     "Description": "Strikes out a claim",
-    "DisplayOrder": 12,
+    "DisplayOrder": 9,
     "PreConditionState(s)": "AWAITING_CLAIMANT_INTENTION",
-    "PostConditionState": "PROCEEDS_WITH_OFFLINE_JOURNEY",
+    "PostConditionState": "CASE_STRUCK_OUT",
     "SecurityClassification": "Public",
     "CallBackURLAboutToStartEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/cases/callbacks/about-to-start",
     "CallBackURLAboutToSubmitEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/cases/callbacks/about-to-submit",
