+++ ccd-definition/CaseField.json
@@ -740,7 +740,7 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "UNSPECIFIED_CLAIMS",
-    "ID": "applicantSolicitorSecondResponseDeadlineToRespondentSolicitor1",
+    "ID": "applicantSolicitor1ClaimStrikeOutDeadlineToRespondentSolicitor1",
     "Label": "Case strikeout deadline",
     "FieldType": "DateTime",
     "SecurityClassification": "Public"
