+++ ccd-definition/State.json
@@ -46,5 +46,13 @@
     "Name": "Proceeds with offline journey",
     "Description": "Case is to be handled offline",
     "DisplayOrder": 6
+  },
+  {
+    "LiveFrom": "01/01/2017",
+    "CaseTypeID": "UNSPECIFIED_CLAIMS",
+    "ID": "CASE_STRUCK_OUT",
+    "Name": "Case struck out",
+    "Description": "No progressive activity on case for 3 months",
+    "DisplayOrder": 7
   }
 ]
