--- src/main/java/uk/gov/hmcts/reform/unspec/handler/tasks/CaseStayedHandler.java
@@ -1,30 +1,0 @@
-package uk.gov.hmcts.reform.unspec.handler.tasks;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.camunda.bpm.client.task.ExternalTask;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.unspec.event.MoveCaseToStayedEvent;
-import uk.gov.hmcts.reform.unspec.service.search.CaseStayedSearchService;
-
-import java.util.List;
-
-@Slf4j
-@RequiredArgsConstructor
-@Component
-public class CaseStayedHandler implements BaseExternalTaskHandler {
-
-    private final CaseStayedSearchService caseSearchService;
-    private final ApplicationEventPublisher applicationEventPublisher;
-
-    @Override
-    public void handleTask(ExternalTask externalTask) {
-        List<CaseDetails> cases = caseSearchService.getCases();
-        log.info("Job '{}' found {} case(s)", externalTask.getTopicName(), cases.size());
-
-        cases.forEach(caseDetails -> applicationEventPublisher.publishEvent(
-            new MoveCaseToStayedEvent(caseDetails.getId())));
-    }
-}
