--- src/test/java/uk/gov/hmcts/reform/unspec/handler/tasks/CaseStayedHandlerTest.java
@@ -1,115 +1,0 @@
-package uk.gov.hmcts.reform.unspec.handler.tasks;
-
-import org.camunda.bpm.client.exception.NotFoundException;
-import org.camunda.bpm.client.task.ExternalTask;
-import org.camunda.bpm.client.task.ExternalTaskService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.unspec.event.MoveCaseToStayedEvent;
-import uk.gov.hmcts.reform.unspec.service.search.CaseStayedSearchService;
-
-import java.util.List;
-import java.util.Map;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
-import static org.mockito.Mockito.when;
-
-@ExtendWith(SpringExtension.class)
-class CaseStayedHandlerTest {
-
-    @Mock
-    private ExternalTask mockTask;
-
-    @Mock
-    private ExternalTaskService externalTaskService;
-
-    @Mock
-    private CaseStayedSearchService searchService;
-
-    @Mock
-    private ApplicationEventPublisher applicationEventPublisher;
-
-    @InjectMocks
-    private CaseStayedHandler caseStayedFinder;
-
-    @BeforeEach
-    void init() {
-        when(mockTask.getTopicName()).thenReturn("test");
-        when(mockTask.getWorkerId()).thenReturn("worker");
-    }
-
-    @Test
-    void shouldEmitMoveCaseToStayedEvent_whenCasesFound() {
-        long caseId = 1L;
-        Map<String, Object> data = Map.of("data", "some data");
-        List<CaseDetails> caseDetails = List.of(CaseDetails.builder().id(caseId).data(data).build());
-
-        when(searchService.getCases()).thenReturn(caseDetails);
-
-        caseStayedFinder.execute(mockTask, externalTaskService);
-
-        verify(applicationEventPublisher).publishEvent(new MoveCaseToStayedEvent(caseId));
-        verify(externalTaskService).complete(mockTask);
-    }
-
-    @Test
-    void shouldNotEmitMoveCaseToStayedEvent_WhenNoCasesFound() {
-        when(searchService.getCases()).thenReturn(List.of());
-
-        caseStayedFinder.execute(mockTask, externalTaskService);
-
-        verifyNoInteractions(applicationEventPublisher);
-    }
-
-    @Test
-    void shouldCallHandleFailureMethod_whenExceptionFromBusinessLogic() {
-        String errorMessage = "there was an error";
-
-        when(mockTask.getRetries()).thenReturn(null);
-        when(searchService.getCases()).thenAnswer(invocation -> {
-            throw new Exception(errorMessage);
-        });
-
-        caseStayedFinder.execute(mockTask, externalTaskService);
-
-        verify(externalTaskService, never()).complete(mockTask);
-        verify(externalTaskService).handleFailure(
-            eq(mockTask),
-            eq(errorMessage),
-            anyString(),
-            eq(2),
-            eq(500L)
-        );
-    }
-
-    @Test
-    void shouldNotCallHandleFailureMethod_whenExceptionOnCompleteCall() {
-        String errorMessage = "there was an error";
-
-        doThrow(new NotFoundException(errorMessage)).when(externalTaskService).complete(mockTask);
-
-        caseStayedFinder.execute(mockTask, externalTaskService);
-
-        verify(externalTaskService, never()).handleFailure(
-            any(ExternalTask.class),
-            anyString(),
-            anyString(),
-            anyInt(),
-            anyLong()
-        );
-    }
-}
